+++ src/game/tools/scriptpack/CMakeLists.txt
@@ -33,6 +33,7 @@ set(SRC_Files
 	"dirent.h"
 	"packer.cpp"
 	"packer.h"
+	"parser.cpp"
 	"parser.h"
 	"scriptpack.cpp"
 	"${BASEDIR}/src/game/shared/ms/crc/checksum_crc.cpp"
